llm/requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
SQLAlchemy
psycopg2-binary
python-dotenv
